Module temple::game::attributes

Attribute definitions and systems.

About Attributes

Attributes are Temple’s translation of a Bevy component into a manifest-defined space for use with sprites (see crate::sprite). In most scenario’s, they are simply tags to apply a system, but attributes additionally have the ability to take parameters to generalize their functionality.

Re-exports

pub use checkpoint::*;

pub use deadly::*;

pub use moving::*;

pub use player::*;

pub use solid::*;

Modules

checkpoint

Sets the players checkpoint. checkpoint(optional x offset, optional y offset)

deadly

Makes sprite deadly to the player. deadly()

moving

Defines a cyclically moving sprite. moving(dir, distance, time)

player

Represents the player. player()

solid

Makes sprite solid to the player. solid()

Structs

AttributeEntry

Attribute with supplied parameters.

AttributePlugin

[Plugin] for attributes

Traits

Attribute

Attribute, as used with a crate::sprite::SpriteType

Functions

build_attribute

Constructs attribute onto a given [Entity]. Used during level load (see crate::level::load::load_level).

derive_attribute_entry

Parses attribute string as a key and supplied parameters.