Module temple::game::attributes::moving

Defines a cyclically moving sprite. moving(dir, distance, time)

TODO: speed should be changed to dist, for easier usage.

Usage

dir: Direction of movement

0 Right

1 Right

2 Right

3 Right

speed: Max velocity of the moving sprite

dur: Duration of the sprite’s cycle in seconds

Structs

MovingSprite

moving attribute state.

Enums

MovingAttributeSystemSteps

Simulation steps for moving sprites. Used for applying force to the player when riding a moving sprite. (see super::AttributePlugin)

MovingDirection

Direction of sprite movement.

Functions

move_player

Moves the Player if they are on top of a moving sprite. TODO: Make it move all entities with a Movable Attribute instead of just the player.

moving_system

System to move all moving sprites per change in [Time].