Module temple::game::player

Player attribute systems

TODO: This plugin should be moved into the same file as the Player attribute.

Structs

PlayerInputCommands

A command queue to handle the revoking, granting, and override of player inputs. Handled as a queue for the later usage of override input, which may require the chaining of override input commands for a cutscene.

PlayerPlugin

[Plugin] for player systems.

Enums

PlayerInputCommand

Constants

PLAYER_JUMP_FORCE

PLAYER_MOVE_SPEED

Functions

handle_dash

handle_dash_crosshair

handle_height_adjust

Consumes [Kurinji] inputs for player hover height adjustments.

handle_player_hover

Handles player hovering.

handle_player_input_commands

handle_player_jump

Consumes [Kurinji] inputs for player jumping.

handle_player_movement

Consumes [Kurinji] inputs for player horizontal movement.

handle_player_slow_fall

Handles Player Slow Falling. When JUMP is actively being held, the player falls slower.