Module temple::level::config

Handles all config manifest loading for levels.

Level manifests are loaded on game boot and are stored in a Bevy Resource.

Structs

HandledSprite

Sprite with loaded texture, to be used in game.

JoinedLevelSpriteEntry

LevelSpriteEntry joined with position from [LevelMapSprite]

Level

Stored object of a level, stores associated info and list of sprites used in level

LevelManifest

Structure of a level toml file.

LevelMap

Map of Level, stored as a binary file in levelmaps/

LevelMapSpriteEntry

Sprite definitions for a level

LevelSpriteEntry

Sprites used in a Level, defined in a LevelManifest

Constants

SPRITE_SIZE

Conversion of game unit to pixel.