Module temple::level::load

Handles level loading for gameplay.

Usage

Levels are instructed to load using tag LoadLevel. LevelLoadComplete can be tracked for completion of load. All sprites that are loaded via LoadLevel are tagged with LevelLoadedSprite. Instruction UnloadLevel can be added to the original LoadLevel entity to instruct an unload.

Structs

LevelLoadComplete
Tag that LoadLevel has completed. Added to same entity as LoadLevel

LevelLoadedSprite
Tag that entity was loaded by level, and will be removed when UnloadLevel instruction is given

LoadLevel
Instruction to load a new level

UnloadLevel
Instruction to unload a level. Must be added to the same entity as LoadLevel

Functions

configure_rapier
Startup system to configure rapier physics for sprites

load_level
System that loads sprites in a given level. Can be tracked with LevelLoadComplete

unload_level
System that unloads a currently loaded level using the UnloadLevel tag