Module temple::sprite

Sprite and sprite type config management

Usage

Sprite types and sprites are loaded on game boot. The textures supplied by the manifest's SpriteEntry list are loaded into Bevy's [AssetServer] for later use, and then are converted to a GameSprite with a reference to the loaded texture and associated attributes.

Structs

GameSprite
Sprite with loaded texture, to be used in game.

SpriteEntry
Sprite Object within sprites/sprites.toml

SpriteFile
Object of sprites/sprites.toml

SpriteFileVersion
Global sprite file version

SpritePlugin
[Plugin] for handling loading sprite config files (attribute and sprite definitions)

SpriteType
Type Object within sprites/types.toml

SpriteTypesFile
Object of sprites/types.toml

Enums

SpritePluginSteps
System labels for sprite loading steps.

Constants

SPRITE_SIZE
Conversion of game unit to pixel.

Functions

load_sprite_types
Startup system that loads all sprite types into memory map

load_sprites
Startup system that loads all sprites and sprite textures into memory

Type Definitions

SpriteId
Sprite ID

SpriteMap
Map that stores loaded sprites.

SpriteTypeMap
Map that stores loaded sprite types.