Module temple::util::files

Handles manifest validation

About

Given temple’s reliance on configurability via manifests, steps should be taken to ensure all config files are in a valid state before game start.

This is currently handled by the entrypoint verify_files, which:

Loads all files from their default locations.

Parses them from their TOML string.

Reports any found issues.

Constants

GAME_SETTING_PATH

game.toml location

LEVEL_FILE_PATH

levels.toml location

SPRITE_FILE_PATH

sprites.toml location

SPRITE_TYPE_FILE_PATH

types.toml location

Functions

find_toml_problems

Consumes the TOML result and reports any errors found.

from_game_root

verify_files

Verifies all game config files are found and valid.