Module temple::game::attributes

Attribute definitions and systems.

About Attributes

Attributes are Temple’s translation of a Bevy component into a manifest-defined space for use with sprites (see [crate::sprite]). In most scenario’s, they are simply tags to apply a system, but attributes additionally have the ability to take parameters to generalize their functionality.

Re-exports

pub use checkpoint::*;

pub use dash::*;

pub use deadly::*;

pub use give::*;

pub use goal::*;

pub use moving::*;

pub use player::*;

pub use solid::*;

pub use transition::*;

Modules

checkpoint

Sets the players checkpoint. checkpoint(optional x offset, optional y offset)

dash

Grants the player a dash. dash()

deadly

Makes sprite deadly to the player. deadly()

give

Gives a player the component provided within the parameter. give(component_name(params...))

goal

Completes an exit for the active level. goal(exit number)

lex

moving

Defines a cyclically moving sprite. moving(dir, distance, time)

player

Represents the player. player()

solid

Makes sprite solid to the player. solid()

transition

Transitions the player to a new level. trans(new level id)

Macros

attribute_build_tree

Structs

AttributePlugin

[Plugin] for attributes

Traits

Attribute

Attribute, as used with a [crate::sprite::SpriteType]

Functions

build_attribute

Constructs attribute onto a given [Entity]. Used during level load (see crate::level::load::load_level).