Module temple::level::load

Handles level loading for gameplay.

Usage

Levels are instructed to load using tag LoadLevel. LevelLoadComplete can be tracked for completion of load. All sprites that are loaded via LoadLevel are tagged with LevelLoadedSprite. Instruction UnloadLevel can be added to the original LoadLevel entity to instruct an unload.

Structs

KeepMusic

Instruction used by transition level to prevent the same music from being replayed.

LevelLoadComplete

Tag that LoadLevel has completed. Added to same entity as LoadLevel

LevelLoadedSprite

Tag that entity was loaded by level, and will be removed when UnloadLevel instruction is given

LevelSaveApplied

Tag to track a level having save files applied to it.

LoadLevel

Instruction to load a new level

PreparedLevel

Loaded level manifests, as part of the load level process

TransitionLevel

Instruction to unload current level and transition to another.

UnloadLevel

Instruction to unload a level. Must be added to the same entity as LoadLevel

WaitUntilUnloaded

Instruction to not LoadLevel until previous level is unloaded.

Functions

apply_save_on_load

Applies the checkpoint location if an active save warrants it.

load_level

System that loads sprites in a given level. Can be tracked with LevelLoadComplete

prepare_level

System that prepares the level from files, to be loaded by load_level

transition_level

System that consumes TransitionLevel instructions Produces

unload_level

System that unloads a currently loaded level using the UnloadLevel tag

wait_until_unloaded